test: move xsnap/snapstore integration test to solo
to avoid @agoric/xsnap in swing-store-lmdb devDependencies
